Factor out auth logic to AuthService
